Delete a database

+++ 0x0D-SQL_introduction/2-remove_database.sql
@@ -1,2 +1,2 @@
 -- Deletes database 'hbtn_0c_0' in MySQL Server
-DROP DATABASE hbtn_0c_0;
+IF EXISTS AND NOT EXISTS DROP DATABASE hbtn_0c_0;
